refactor: split departments controller into crud/health/ceremony_policy package per ADR-0008 (#2253)
